Split test files, all tests pass except simplify, began adding simplify functionality
