+++ sympy/tensor/algebraic/__init__.py
@@ -3,6 +3,7 @@ from sympy.tensor.algebraic.algebraic_tensor import (
     ShapeMismatchError,
 )
 from sympy.tensor.algebraic.pure_tensor import PureTensor, tensor_product
+from sympy.tensor.algebraic.simplify import tensorsimplify
 from sympy.tensor.algebraic.zero_tensor import ZeroTensor, zero_tensor
 
 __all__ = [
@@ -11,5 +12,6 @@ __all__ = [
     "ShapeMismatchError",
     "ZeroTensor",
     "tensor_product",
+    "tensorsimplify",
     "zero_tensor",
 ]
